Load the source of countyLayer manually
Currently county styles are applied on a per state basis
when a state is hovered.
But because of lower layer maxResolution the county level features
aren't available until we zoom in far enough to display it.
We fix this by manually fetching the TopoJSON
and directly adding the features.
A hacky alternative is to increase maxResolution when layer is defined,
then at some point (e.g. when stateLayer is loaded (changed+ready state)
lowering it back with setMaxResolution().