Fix county level fill bug
County layer features are styled directly if they contain data.
There is a default style for the county layer which applies to the
remaining county features, but this shares a global style object
with the state layer.
Because the state layer style function changes the fill on this object,
the county layer fill style would change erratically.