Store fillColor on state features and use it to style
Previously the story_count was fetched from preview and color computed
in the style function, which was defined after the previews were loaded
(in fetch then clause). Attaching the fillColor directly to features
has potential performance and other benefits. However, it also means
that the features must actually exist when this is done, however that
was not necessary when doing it all in the style function  since the
feature was not accessed until the style function was executed.