Create getStatePreviews() function to delay loading
The previous commit made it necessary for stateLayer features to be
loaded before the state-level preview since properties are set on the
features immediately after preview is loaded. To ensure this the loading
of the state-level preview is wrapped in a function (vs on script run)
which is not called until after the stateLevel features are loaded.